+++ src/01_childOp.js
@@ -1,6 +1,6 @@
 let U = require('./util')
 const c = require('./constants');
-const Operation = require('./000_operation');
+const Operation = require('./00_operation');
 
 module.exports = class ChildOp extends Operation{
     /**@param {Operation} parent */
