+++ src/11_shardChildOp.js
@@ -1,7 +1,7 @@
 const U = require('./util');
 const c = require('./constants');
-const ChildOp = require('./childOp');
-const CreepOp = require('./creepOp');
+const ChildOp = require('./01_childOp');
+const CreepOp = require('./12_creepOp');
 
 module.exports = class ShardChildOp extends ChildOp {
     /**
