+++ src/22_harvestingOp.js
@@ -1,6 +1,6 @@
 const U = require('./util');
 const c = require('./constants');
-const BaseChildOp = require('./baseChildOp');
+const BaseChildOp = require('./21_baseChildOp');
 
 const HARVESTER_SIZE_BIG = 48
 const HARVESTER_SIZE_SMALL = 6*3
