+++ src/22_towerOp.js
@@ -1,6 +1,6 @@
 let U = require('./util');
 const c = require('./constants');
-const BaseChildOp = require('./baseChildOp');
+const BaseChildOp = require('./21_baseChildOp');
 
 const MAX_HITS_REPAIR_PER_LEVEL = 10000
